refactor(plugin): more "de-mixin"-ification, this time the wasm-pack calls
